+++ src-tauri/src/lib.rs
@@ -43,7 +43,8 @@ pub use adapters::driving::tauri_ipc::{
 pub fn run() {
     let mut builder = tauri::Builder::default()
         .plugin(tauri_plugin_clipboard_manager::init())
-        .plugin(tauri_plugin_notification::init());
+        .plugin(tauri_plugin_notification::init())
+        .plugin(tauri_plugin_updater::Builder::new().build());
     #[cfg(all(debug_assertions, unix))]
     {
         builder = builder.plugin(tauri_plugin_pilot::init());